everything working nicely on grayscale pics

+++ jpeglib.h
@@ -855,10 +855,10 @@ struct jpeg_memory_mgr {
    * used for virtual-array buffers.)  May be changed by outer application
    * after creating the JPEG object.
    */
-  long max_memory_to_use;
+  long long max_memory_to_use;
 
   /* Maximum allocation request accepted by alloc_large. */
-  long max_alloc_chunk;
+  long long max_alloc_chunk;
 };
